+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Foundry 0.8.3 support
 * Integrated adventure importer tool
+* You can now specify an Override compendium to import character items/spells/features from.
+* New only option has been removed.
 
 # 1.0.9